+++ eulerpi/examples/heat/__init__.py
@@ -1,0 +1,2 @@
+from .heat import Heat as Heat
+from .heat import HeatArtificial as HeatArtificial